Moves existing screens to a subpackage
