+++ app/src/main/java/com/adriencadet/wanderer/ui/screens/app/PlaceInsightScreen.java
@@ -1,4 +1,4 @@
-package com.adriencadet.wanderer.ui.screens;
+package com.adriencadet.wanderer.ui.screens.app;
 
 import com.adriencadet.wanderer.models.bll.dto.PlaceBLLDTO;
 import com.adriencadet.wanderer.ui.controllers.PlaceInsightController;
