+++ app/src/main/java/com/adriencadet/wanderer/ui/screens/popup/AlertScreen.java
@@ -1,0 +1,8 @@
+package com.adriencadet.wanderer.ui.screens.popup;
+
+/**
+ * AlertScreen
+ * <p>
+ */
+public class AlertScreen {
+}
